Заработало появление активного маркера на карте
